test: refatoração dos testes
